requirements.txt
robotframework
robotframework-seleniumlibrary
webdriver-manager
google-auth 
google-auth-oauthlib 
google-api-python-client 
beautifulsoup4
robotframework-databaselibrary 
psycopg2-binary